SlowTower

+++ 2023_2_Project_A/Assets/Terasurware/Classes/Editor/GameDB_importer.cs
@@ -50,7 +50,7 @@ public class GameDB_importer : AssetPostprocessor {
 						
 						Entity_GameDB.Param p = new Entity_GameDB.Param ();
 						
-					cell = row.GetCell(0); p.index = (cell == null ? 0.0 : cell.NumericCellValue);
+					cell = row.GetCell(0); p.index = (int)(cell == null ? 0 : cell.NumericCellValue);
 					cell = row.GetCell(1); p.characterName = (cell == null ? "" : cell.StringCellValue);
 					cell = row.GetCell(2); p.hp = (int)(cell == null ? 0 : cell.NumericCellValue);
 					cell = row.GetCell(3); p.mp = (int)(cell == null ? 0 : cell.NumericCellValue);
