+++ 2023_2_Project_A/Assets/Terasurware/Classes/Entity_GameDB.cs
@@ -17,7 +17,7 @@ public class Entity_GameDB : ScriptableObject
 	public class Param
 	{
 		
-		public double index;
+		public int index;
 		public string characterName;
 		public int hp;
 		public int mp;
